server/requirements.txt
flask-sandboy
flask-sqlalchemy
mysql-python
